feat: support drift detection for terraform deployer
